replaced nav with simpler header, set up as page template
replaced the nav bar with a simpler header and link bar with a more
uniform look + feel to the login screen. doesn't display which page
you're currently on as any sort of navigation header link highlight:
looking into replacing that functionality however. now project must be
run on tomcat in order to load properly however as its dependent on jsp
inclusions of the header/footer for all pages. see README for more info.